requirements.txt
flask
flask-socketio
pyyaml
eventlet
